Sei
"sei" - a project aimed at creating an innovative platform for exchanging knowledge and experience in the field of artificial intelligence and machine learning. The main goal of the project is to promote the development of artificial intelligence technologies through collaborative learning and collaboration among specialists and researchers from around the world.